+++ plant_water_bot/RHTimer.h
@@ -6,13 +6,13 @@
 class RHTimer {
     public:
         static void start();
-        static void start(int interval);
-        static int get_d_remaining();
-        static int get_h_remaining();
-        static int get_m_remaining();
-        static int get_s_remaining();
+        static void start(uint8_t interval);
+        static uint8_t get_d_remaining();
+        static uint8_t get_h_remaining();
+        static uint8_t get_m_remaining();
+        static uint8_t get_s_remaining();
         static void update();
-        static int get_current_interval();
+        static uint8_t get_current_interval();
 
     private:
